backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0
pyotp==2.9.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI Integration
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.35.2

# File Upload & Processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.1.0

# Email
emails==0.6.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2